+++ HISTORY.md
@@ -4,12 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.11.1]
 ### Removed
 - Deprecated applicationConfigs on register
 
 ### Added
 - Emit app-config to mxss on app start
+- Save device wifi info to firebase
+  
+### Changed
+- Reenabled service.then
 
 ## [0.11.0]
 ### Changed